+++ flutter_inappwebview_windows/windows/in_app_webview/in_app_webview.h
@@ -139,7 +139,7 @@ namespace flutter_inappwebview_plugin
       double x, double y, double size, double pressure);
     void setPointerButtonState(InAppWebViewPointerEventKind kind, InAppWebViewPointerButton button);
     void sendScroll(double offset, bool horizontal);
-    void setScrollDelta(double delta_x, double delta_y);
+    void setScrollDelta(double delta_x, double delta_y, bool isPan = false);
     void onSurfaceSizeChanged(SurfaceSizeChangedCallback callback)
     {
       surfaceSizeChangedCallback_ = std::move(callback);
@@ -222,20 +222,20 @@ namespace flutter_inappwebview_plugin
     void closeWebMessagePort(const std::string& channelId, int portIndex);
     void disposeWebMessageChannel(const std::string& channelId);
 
-    void addWebNotificationController(const std::string& id, std::shared_ptr<WebNotificationController> controller);
-    WebNotificationController* getWebNotificationController(const std::string& id) const;
-    void eraseWebNotificationController(const std::string& id);
+    void addWebNotificationController(const std::string& controllerId, std::shared_ptr<WebNotificationController> controller);
+    WebNotificationController* getWebNotificationController(const std::string& controllerId) const;
+    void eraseWebNotificationController(const std::string& controllerId);
     void disposeAllWebNotificationControllers();
 
-    void addPrintJobController(const std::string& id, std::shared_ptr<PrintJobController> controller);
-    PrintJobController* getPrintJobController(const std::string& id) const;
-    void erasePrintJobController(const std::string& id);
+    void addPrintJobController(const std::string& controllerId, std::shared_ptr<PrintJobController> controller);
+    PrintJobController* getPrintJobController(const std::string& controllerId) const;
+    void erasePrintJobController(const std::string& controllerId);
     void disposeAllPrintJobControllers();
 
-    void printCurrentPage(std::shared_ptr<PrintJobSettings> settings,
+    void printCurrentPage(std::shared_ptr<PrintJobSettings> jobSettings,
       const std::function<void(const std::optional<std::string>&)> completionHandler);
 
-    void createPdf(std::shared_ptr<PrintJobSettings> settings,
+    void createPdf(std::shared_ptr<PrintJobSettings> jobSettings,
       const std::function<void(const std::optional<std::vector<uint8_t>>&)> completionHandler);
 
     std::string pageFrameId() const
@@ -252,6 +252,9 @@ namespace flutter_inappwebview_plugin
     float scaleFactor_ = 1.0;
     POINT lastCursorPos_ = { 0, 0 };
     VirtualKeyState virtualKeys_;
+    double scrollDeltaAccumulatorX_ = 0.0;
+    double scrollDeltaAccumulatorY_ = 0.0;
+    double flutterToWheelDeltaScale_ = 1.2; // WHEEL_DELTA * 3 / (lines * 100)
 
     const std::string expectedBridgeSecret = get_uuid();
     bool javaScriptBridgeEnabled = true;